requirements.txt
rdkit>=2023.9.1
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
pinecone>=5.0.0
opensearch-py>=2.6.0
langchain-google-genai>=0.1.0
langchain-anthropic>=1.0.0
boto3>=1.34.0
numpy>=1.24.0
reportlab>=4.4.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.18.0
pytest>=7.0.0
hypothesis>=6.0.0
psutil>=5.9.0
docker>=6.0.0
requests>=2.28.0
tenacity>=8.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
slowapi>=0.1.9
stmol>=0.0.9
py3Dmol>=2.0.0
ipython_genutils>=0.2.0
streamlit-lottie>=0.0.5

# Open-source embeddings and vector DB alternatives
sentence-transformers>=2.2.0
qdrant-client>=1.7.0

# Optional VCF streaming cache
redis>=5.0.0

# FHIR push client
httpx>=0.27.0

# AWS Competition Enhancement Dependencies
diagrams>=0.23.0
graphviz>=0.20.0
matplotlib>=3.7.0
seaborn>=0.12.0
langchain_anthropic>=0.1.0
